requirements.txt
black~=23.1
gTTS~=2.3
openai~=0.28
SpeechRecognition~=3.10
pyaudio~=0.2
pydub~=0.25
pyngrok~=6.1.0
gevent~=23.9.1
twilio~=8.8.0
flask~=2.3
flask_sock~=0.6
sqlalchemy~=2.0
flask-sqlalchemy~=3.0
python-dateutil~=2.8
groq~=0.4.1
python-dotenv~=1.0.0
